Move LoginViewModel back to ui directory.
