+++ libs/SalesforceSDK/src/com/salesforce/androidsdk/ui/components/LoginServerBottomSheet.kt
@@ -31,7 +31,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.salesforce.androidsdk.app.SalesforceSDKManager
-import com.salesforce.androidsdk.auth.LoginViewModel
+import com.salesforce.androidsdk.ui.LoginViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
