+++ libs/SalesforceSDK/src/com/salesforce/androidsdk/ui/components/LoginServerCard.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.salesforce.androidsdk.app.SalesforceSDKManager
 import com.salesforce.androidsdk.config.LoginServerManager
-import com.salesforce.androidsdk.auth.LoginViewModel
+import com.salesforce.androidsdk.ui.LoginViewModel
 
 //    @Preview
 @Composable
